chore: extract poller handler
