+++ packages/client/hmi-client/src/components/workflow/tera-calibrate-ensemble-node-ciemss.vue
@@ -55,10 +55,10 @@ import {
 import { workflowEventBus } from '@/services/workflow';
 import { EnsembleCalibrationCiemssRequest, TimeSpan, EnsembleModelConfigs } from '@/types/Types';
 import {
-	getSimulation,
 	makeEnsembleCiemssCalibration,
 	getRunResultCiemss,
-	handleSimulationsInProgress
+	handleSimulationsInProgress,
+	simulationPollAction
 } from '@/services/models/simulation-service';
 import Button from 'primevue/button';
 import { ChartConfig, RunResults } from '@/types/SimulateConfig';
@@ -133,62 +133,11 @@ const getStatus = async (simulationId: string) => {
 	showSpinner.value = true;
 	if (!simulationId) return;
 
+	const runIds = [simulationId];
 	const poller = new Poller<object>()
 		.setInterval(3000)
 		.setThreshold(300)
-		.setPollAction(async () => {
-			const response = await getSimulation(simulationId);
-			if (response?.status === ProgressState.COMPLETE) {
-				const newState = handleSimulationsInProgress(
-					SimulationStateOperation.DELETE,
-					props.node,
-					simulationId
-				);
-				if (newState) {
-					emit('update-state', newState);
-				}
-				return {
-					data: response,
-					progress: null,
-					error: null
-				};
-			}
-			if (response?.status === ProgressState.RUNNING) {
-				const newState = handleSimulationsInProgress(
-					SimulationStateOperation.ADD,
-					props.node,
-					simulationId
-				);
-				if (newState) {
-					emit('update-state', newState);
-				}
-				progress.value = {
-					status: ProgressState.RUNNING,
-					value: 0
-				};
-			}
-
-			if (response?.status === ProgressState.QUEUED) {
-				const newState = handleSimulationsInProgress(
-					SimulationStateOperation.ADD,
-					props.node,
-					simulationId
-				);
-				if (newState) {
-					emit('update-state', newState);
-				}
-				progress.value = {
-					status: ProgressState.QUEUED,
-					value: 0
-				};
-			}
-
-			return {
-				data: null,
-				progress: null,
-				error: null
-			};
-		});
+		.setPollAction(async () => simulationPollAction(runIds, props.node, progress, emit));
 	const pollerResults = await poller.start();
 
 	if (pollerResults.state !== PollerState.Done || !pollerResults.data) {
