+++ packages/client/hmi-client/src/components/workflow/tera-calibration-node-julia.vue
@@ -116,9 +116,9 @@ import AccordionTab from 'primevue/accordiontab';
 import { CalibrationRequestJulia, CsvAsset, ModelConfiguration, TimeSpan } from '@/types/Types';
 import {
 	makeCalibrateJobJulia,
-	getSimulation,
 	getRunResultJulia,
-	handleSimulationsInProgress
+	handleSimulationsInProgress,
+	simulationPollAction
 } from '@/services/models/simulation-service';
 import { setupModelInput, setupDatasetInput } from '@/services/calibrate-workflow';
 import { ChartConfig, RunResults } from '@/types/SimulateConfig';
@@ -235,62 +235,11 @@ const getStatus = async (simulationId: string) => {
 	showSpinner.value = true;
 	if (!simulationId) return;
 
+	const runIds = [simulationId];
 	const poller = new Poller<object>()
 		.setInterval(3000)
 		.setThreshold(300)
-		.setPollAction(async () => {
-			const response = await getSimulation(simulationId);
-			if (response?.status === ProgressState.COMPLETE) {
-				const newState = handleSimulationsInProgress(
-					SimulationStateOperation.DELETE,
-					props.node,
-					simulationId
-				);
-				if (newState) {
-					emit('update-state', newState);
-				}
-				return {
-					data: response,
-					progress: null,
-					error: null
-				};
-			}
-			if (response?.status === ProgressState.RUNNING) {
-				const newState = handleSimulationsInProgress(
-					SimulationStateOperation.ADD,
-					props.node,
-					simulationId
-				);
-				if (newState) {
-					emit('update-state', newState);
-				}
-				progress.value = {
-					status: ProgressState.RUNNING,
-					value: 0
-				};
-			}
-
-			if (response?.status === ProgressState.QUEUED) {
-				const newState = handleSimulationsInProgress(
-					SimulationStateOperation.ADD,
-					props.node,
-					simulationId
-				);
-				if (newState) {
-					emit('update-state', newState);
-				}
-				progress.value = {
-					status: ProgressState.QUEUED,
-					value: 0
-				};
-			}
-
-			return {
-				data: null,
-				progress: null,
-				error: null
-			};
-		});
+		.setPollAction(async () => simulationPollAction(runIds, props.node, progress, emit));
 	const pollerResults = await poller.start();
 
 	if (pollerResults.state !== PollerState.Done || !pollerResults.data) {
