+++ packages/client/hmi-client/src/components/workflow/tera-simulate-ciemss-node.vue
@@ -46,15 +46,14 @@ import AccordionTab from 'primevue/accordiontab';
 import Dropdown from 'primevue/dropdown';
 import {
 	makeForecastJobCiemss as makeForecastJob,
-	getSimulation,
 	getRunResultCiemss,
-	handleSimulationsInProgress
+	handleSimulationsInProgress,
+	simulationPollAction
 } from '@/services/models/simulation-service';
 import InputNumber from 'primevue/inputnumber';
 import { ProgressState, SimulationStateOperation, WorkflowNode } from '@/types/workflow';
 import { ChartConfig, RunResults } from '@/types/SimulateConfig';
 import { workflowEventBus } from '@/services/workflow';
-import { Simulation } from '@/types/Types';
 import { Poller, PollerState } from '@/api/api';
 import TeraSimulateChart from './tera-simulate-chart.vue';
 import { SimulateCiemssOperation, SimulateCiemssOperationState } from './simulate-ciemss-operation';
@@ -111,68 +110,21 @@ onMounted(() => {
 	}
 });
 
-const getStatus = async (simulationIds: string[]) => {
+const getStatus = async (runIds: string[]) => {
 	showSpinner.value = true;
 	const poller = new Poller<object>()
 		.setInterval(3000)
 		.setThreshold(300)
-		.setPollAction(async () => {
-			const requestList: Promise<Simulation | null>[] = [];
-			simulationIds.forEach((id) => {
-				requestList.push(getSimulation(id));
-			});
-			const response = await Promise.all(requestList);
-
-			if (response.every((simulation) => simulation!.status === ProgressState.COMPLETE)) {
-				const newState = handleSimulationsInProgress(
-					SimulationStateOperation.DELETE,
-					props.node,
-					simulationIds
-				);
-				if (newState) {
-					emit('update-state', newState);
-				}
-				return {
-					data: response,
-					progress: null,
-					error: null
-				};
-			}
-			if (
-				response.find(
-					(simulation) =>
-						simulation?.status === ProgressState.QUEUED ||
-						simulation?.status === ProgressState.RUNNING
-				)
-			) {
-				const newState = handleSimulationsInProgress(
-					SimulationStateOperation.ADD,
-					props.node,
-					simulationIds
-				);
-				progress.value = {
-					status: ProgressState.RUNNING,
-					value: 0
-				};
-				if (newState) {
-					emit('update-state', newState);
-				}
-			}
-			return {
-				data: null,
-				progress: null,
-				error: null
-			};
-		});
+		.setPollAction(async () => simulationPollAction(runIds, props.node, progress, emit));
 	const pollerResults = await poller.start();
 
 	if (pollerResults.state !== PollerState.Done || !pollerResults.data) {
 		// throw if there are any failed runs for now
-		console.error('Failed', simulationIds);
+		console.error('Failed', runIds);
 		showSpinner.value = false;
 		throw Error('Failed Runs');
 	}
-	completedRunIdList.value = simulationIds;
+	completedRunIdList.value = runIds;
 	showSpinner.value = false;
 };
